Rename to make it clear that it's miro specific
