+++ miro_preprocessor/lambdas/sns_to_dynamo/src/miro_image_info_to_dynamo.py
